[API] move code except for endpoints on FE to package
