+++ apps/web/package.json
@@ -24,6 +24,7 @@
     "@mui/material": "5.15.2",
     "@next/bundle-analyzer": "14.0.4",
     "animate-css-grid": "1.5.1",
+    "api": "workspace:*",
     "api-clients": "workspace:*",
     "db": "workspace:*",
     "dotenv-mono": "1.3.13",
