+++ apps/web/src/components/maps/FullMapCard.tsx
@@ -1,11 +1,11 @@
 import { Maximize2, Minimize2 } from 'lucide-react';
 import { useMemo, useState } from 'react';
 import { useTheme } from '@mui/material';
+import type { MapLocation } from 'api/contentful/MapLocation';
 import type { ContentCardProps } from 'components/ContentCard';
 import { Map } from 'components/maps/Map';
 import { Marker } from 'components/maps/Marker';
 import { Control } from 'components/maps/Control';
-import type { MapLocation } from 'api/server/contentful/MapLocation';
 import { MapContentCard } from './MapContentCard';
 
 type FullMapCardProps = Pick<ContentCardProps, 'turnOnAnimation'> & {
