+++ apps/web/src/components/spotify/ArtistList.tsx
@@ -4,7 +4,7 @@
 import { Fragment } from 'react';
 import { Typography } from '@mui/material';
 import { truncated } from 'ui/helpers/truncated';
-import type { Artist } from 'api/server/spotify/Track';
+import type { Artist } from 'api/spotify/Track';
 import { Link } from 'components/Link';
 import { useLinkWithName } from 'hooks/useLinkWithName';
