+++ packages/api/contentful/fetchCurrentLocation.generated.ts
