+++ packages/api/contentful/fetchCurrentLocation.ts
