+++ packages/api/contentful/fetchFooterLinks.generated.ts
