+++ packages/api/contentful/fetchFooterLinks.ts
