+++ packages/eslint-config-dg/library.js
@@ -14,9 +14,12 @@ const project = resolve(process.cwd(), 'tsconfig.json');
 
 /** @type {import('eslint').Linter.Config} */
 module.exports = {
-  extends: ['@vercel/style-guide/eslint/browser', '@vercel/style-guide/eslint/typescript'].map(
-    (importPath) => require.resolve(importPath),
-  ),
+  extends: [
+    '@vercel/style-guide/eslint/node',
+    '@vercel/style-guide/eslint/browser',
+    '@vercel/style-guide/eslint/typescript',
+    'eslint-config-turbo',
+  ].map((importPath) => require.resolve(importPath)),
   parser: '@typescript-eslint/parser',
   parserOptions: {
     project,
