+++ turbo.json
@@ -21,22 +21,23 @@
   ],
   "globalEnv": [
     "ANALYZE",
-    "NEXT_PUBLIC_MAPBOX_TOKEN",
-    "CONTENTFUL_SPACE_ID",
     "CONTENTFUL_ACCESS_TOKEN",
-    "GITHUB_AUTHENTICATION_TOKEN",
+    "CONTENTFUL_SPACE_ID",
     "CONTENTFUL_SPACE_ID",
     "DATABASE_URL",
-    "WEBHOOK_CALLBACK_URL",
+    "GITHUB_AUTHENTICATION_TOKEN",
+    "NEXT_PUBLIC_APP_VERSION",
+    "NEXT_PUBLIC_MAPBOX_TOKEN",
+    "NEXT_PUBLIC_SENTRY_DSN",
+    "NODE_ENV",
+    "SENTRY_AUTH_TOKEN",
+    "SPOTIFY_CLIENT_ID",
+    "SPOTIFY_CLIENT_SECRET",
     "STRAVA_CLIENT_ID",
     "STRAVA_CLIENT_SECRET",
     "STRAVA_TOKEN_NAME",
     "STRAVA_VERIFY_TOKEN",
-    "SPOTIFY_CLIENT_ID",
-    "SPOTIFY_CLIENT_SECRET",
-    "SENTRY_AUTH_TOKEN",
-    "NEXT_PUBLIC_SENTRY_DSN",
-    "NEXT_PUBLIC_APP_VERSION"
+    "WEBHOOK_CALLBACK_URL"
   ],
   "pipeline": {
     "topo": {
